MobileMechanicCloud/requirements.txt
Flask==0.12.1
Flask-HTTPAuth==3.2.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.5
facebook-sdk==2.0.0
PyMongo==3.4.0
flask-pymongo==0.4.1